product_analyst_multimodal_agent/requirements.txt
streamlit
pillow
phidata
google-generativeai
tavily-python
python-dotenv